Use SlottingParameters from primitive lib in wallet
