+++ lib/primitive/cardano-wallet-primitive.cabal
@@ -85,6 +85,7 @@ library
     , scientific
     , text
     , text-class
+    , time
     , transformers
     , unliftio
     , unliftio-core
@@ -108,6 +109,7 @@ library
     Cardano.Wallet.Primitive.Types.ProtocolParameters
     Cardano.Wallet.Primitive.Types.RewardAccount
     Cardano.Wallet.Primitive.Types.RewardAccount.Gen
+    Cardano.Wallet.Primitive.Types.SlottingParameters
     Cardano.Wallet.Primitive.Types.TokenBundle
     Cardano.Wallet.Primitive.Types.TokenBundle.Gen
     Cardano.Wallet.Primitive.Types.TokenBundleMaxSize
